#3 - Modifiyng logout, reset_password and confirm_password to class-Based view.

+++ medical_prescription/user/views.py
@@ -7,8 +7,7 @@ from django.contrib import auth
 from django.core.mail import send_mail
 from django.contrib import messages
 from django.utils import timezone
-from django.views.generic import FormView
-from django.http import HttpResponse
+from django.views.generic import FormView, View
 
 from .models import HealthProfessional, User, ResetPasswordProfile
 from .forms import (HealthProfessionalForm,
@@ -29,6 +28,7 @@ def register_view(request):
         password = form.cleaned_data.get('password')
         user = User.objects.create_user(email=email, password=password)
         user.save()
+        return render(request, 'message.html', {"message": "usuário registrado"})
     else:
         # Nothing to do.
         pass
@@ -45,7 +45,6 @@ class LoginView(FormView):
         return render(request, self.template_name, {'form': form})
 
     def post(self, request, *args, **kwargs):
-
         form = self.form_class(request.POST)
 
         if form.is_valid():
@@ -53,41 +52,35 @@ class LoginView(FormView):
                 email=form.cleaned_data['email'],
                 password=form.cleaned_data['password'],
             )
-
-            LoginView.user_authentication(request, user)
-
-        else:
-
-            return HttpResponse('Invalid Form')
-
-    def user_authentication(request, user):
-        '''
-        If the user is authenticated this method log the user.
-        '''
-        success_url = '/home'
-
-        if user is not None:
-            if user.is_active:
-                auth.login(request, user)
-                return redirect(str(success_url))
+            if user is not None:
+                return self.user_authentication(request, user)
             else:
-                # do nothing
-                pass
+                return render(request, 'message.html', {"message": "usuário não none"})
         else:
-            return HttpResponse('User is not active')
+            return render(request, self.template_name, {'form': form})
+
+    def user_authentication(self, request, user):
+        if user.is_active:
+            auth.login(request, user)
+            return redirect('/home')
 
 
-def reset_password(request):
+class ResetPasswordView(FormView):
     '''
     Send an e-mail to reset user password.
     '''
+    form_class = ResetPasswordForm
+    template_name = 'reset_password.html'
 
-    # Get form information
-    form = ResetPasswordForm(request.POST or None)
+    def get(self, request, *args, **kwargs):
+        form = self.form_class(initial=self.initial)
+        return render(request, self.template_name, {'form': form})
 
-    if form.is_valid():
-
-        email = form.cleaned_data.get('email')
+    # Get form information.
+    def post(self, request, *args, **kwargs):
+        form = self.form_class(request.POST)
+        if form.is_valid():
+            email = form.cleaned_data.get('email')
 
         # Search the user in database
         try:
@@ -98,25 +91,13 @@ def reset_password(request):
         # Create a hash with the 'salt' and the user e-mail and send the same to the user.
 
         try:
-
-            # Create a random sha1 code.
-            salt = hashlib.sha1(str(random.random()).encode('utf-8')).hexdigest()[:5]
-
-            # Join 'salt' and 'email' to create the activation key.
-            activation_key = hashlib.sha1(str(salt+email).encode('utf‌​-8')).hexdigest()
-
-            # Make a expire parameter for the activation key.
-            key_expires = datetime.datetime.today() + datetime.timedelta(2)
-
-            new_profile = ResetPasswordProfile(user=user,
-                                               activation_key=activation_key,
-                                               key_expires=key_expires)
+            new_profile = self._create_recover_profile(user)
 
             new_profile.save()
 
             # Standar e-mail text.
             email_subject = 'Recuperar senha'
-            email_body = ('Segue sua chave de ativaçãohttp://0.0.0.0:8000/home/reset_confirm/%s.' % activation_key)
+            email_body = ('Segue sua chave de ativação http://0.0.0.0:8000/home/reset_confirm/%s.' % new_profile.activation_key)
 
             send_mail(email_subject,
                       email_body,
@@ -133,54 +114,81 @@ def reset_password(request):
             messages.error(request, 'um email de recuperação de senha já foi enviado para este endereço!')
             return render(request, 'reset_password.html',
                           {"form": form})
-    else:
-        # nothing to do
-        pass
+        else:
+            # nothing to do
+            pass
 
-    return render(request, 'reset_password.html', {"form": form})
+    # return render(request, 'reset_password.html', {"form": form})
+
+    def _create_recover_profile(self, user):
+            email = user.email
+            # Create a random sha1 code.
+            salt = hashlib.sha1(str(random.random()).encode('utf-8')).hexdigest()[:5]
+
+            # Join 'salt' and 'email' to create the activation key.
+            activation_key = hashlib.sha1(str(salt+email).encode('utf‌​-8')).hexdigest()
+
+            # Make a expire parameter for the activation key.
+            key_expires = datetime.datetime.today() + datetime.timedelta(2)
+
+            new_profile = ResetPasswordProfile(user=user,
+                                               activation_key=activation_key,
+                                               key_expires=key_expires)
+
+            return new_profile
 
 
-def confirm_password(request, activation_key):
-    '''
-    Confirm change password user.
-    '''
+class ConfirmPasswordView(FormView):
 
-    form = ConfirmPasswordForm(request.POST or None)
+    form_class = ConfirmPasswordForm
+    template_name = 'password_confirm.html'
 
-    # Get reset object.
-    user_profile = get_object_or_404(ResetPasswordProfile, activation_key=activation_key)
+    def get(self, request, *args, **kwargs):
+        form = self.form_class(initial=self.initial)
+        return render(request, self.template_name, {'form': form})
 
-    user = user_profile.user
+    def post(self, request, activation_key, *args, **kwargs):
 
-    # Case key expires.
-    if(user_profile.key_expires < timezone.now()):
-        # key expires.
-        user.profile.delete()
-        return redirect('/')
-    else:
-        # Nothing to do.
-        pass
+        form = ConfirmPasswordForm(request.POST or None)
 
-    # Change user password and save in database.
-    if(request.method == 'POST'):
+        # Get reset object.
+        user_profile = get_object_or_404(ResetPasswordProfile, activation_key=activation_key)
+
+        user = user_profile.user
+
+        if(self._validate_activation_key(user_profile) and request.method == 'POST'):
+
+            # Change user password and save in database.
+            self._save_user(user, form)
+            user_profile.delete()
+        else:
+            return redirect('/')
+
+        return render(request, 'password_confirm.html', {'form': form})
+
+    def _validate_activation_key(user_profile):
+        # Case key expires.
+        if(user_profile.key_expires < timezone.now()):
+            # key expires.
+            user_profile.delete()
+            return False
+        else:
+            return True
+
+    def _save_user(user, form):
         if(form.is_valid()):
             user.set_password(form.cleaned_data.get('password'))
             user.save()
-            user_profile.delete()
             return redirect('/')
         else:
             # Nothing to do.
             pass
-    else:
-        # Nothing to do.
-        pass
-
-    return render(request, 'password_confirm.html', {'form': form})
 
 
-def logout_view(request):
-    auth.logout(request)
-    return redirect('/home')
+class LogoutView(View):
+    def get(self, request):
+        auth.logout(request)
+        return redirect('/home/login')
 
 
 def register_health_professional(request):
